requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
sqlite3
openai==1.3.7
sentence-transformers==2.2.2
scikit-learn==1.3.2
hdbscan==0.8.33
numpy==1.24.4
pandas==2.1.4
pydantic==2.5.0
python-dotenv==1.0.0
asyncio
logging 